requirements.txt
pytorch==2.3.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.3.1
tqdm==4.66.4
yaml==0.2.5
comet-ml==3.43.2
addict==2.4.0
torchvision==0.18.0